replaced startfield with chroma inverter
Replaced startfield option with chroma inverter option. Chroma can also now be inverted in live mode. What is done is manually swapping Cb and Cr processing.
The pink image is a result of processing Cb as Cr and vice-versa. This can happen if vhs-decode drops fields and thus the order of CbCr transmission swaps.